+++ src/index.js
@@ -1,18 +1,16 @@
 export { compileCommand, renderCompiled } from './compile.js';
-export { lintCommand, lintPath } from './lint.js';
 export {
-  collectSourceFiles,
-  listProfiles,
-  loadCatalog,
-  loadProfile,
-  MAX_SOURCE_FILE_BYTES,
-  readJson,
-  rootPath
-} from './io.js';
+  DEFAULT_CONFIG_FILE,
+  findSuppression,
+  globToRegExp,
+  loadConfig
+} from './config.js';
+export { lintCommand, lintPath } from './lint.js';
 export { selectRules } from './select.js';
 export {
   assertSafeProfileName,
   validateCatalog,
+  validateConfig,
   validatePattern,
   validateProfile,
   validateRule